refactor: move rendering of content out of the BottomSheet common component, make snapPoints a prop
